Improve code structure and fix php-cs-fixer
